Refactor Login form to it's own component
